move the generic serial task to its own module
