+++ dnp3/src/serial/mod.rs
@@ -52,6 +52,7 @@ pub use outstation::*;
 
 mod master;
 mod outstation;
+pub(crate) mod task;
 
 /// State of the serial port
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
